Refactor cli commands.
